Change the interface of the rowUtils
